Code moved App directory to Index directory
